+++ src/transport/mqtt/mod.rs
@@ -888,7 +888,7 @@ pub(crate) mod test {
         Ok(notice)
     }
 
-    pub(crate) fn mock_mqtt_connection<S>(
+    pub(crate) async fn mock_mqtt_connection<S>(
         client: AsyncClient,
         eventloop: EventLoop,
         store: S,
@@ -907,17 +907,22 @@ pub(crate) mod test {
         let volatile = SharedVolatileStore::with_capacity(DEFAULT_VOLATILE_CAPACITY);
 
         let store = StoreWrapper::new(store);
+
+        let transport_provider = TransportProvider::configure(
+            "http://api.astarte.localhost/pairing".parse().unwrap(),
+            "secret".to_string(),
+            None,
+            true,
+        )
+        .await
+        .unwrap();
+
         let mqtt = Mqtt::new(
             client_id.clone(),
             MqttConnection::new(
                 client.clone(),
                 eventloop,
-                TransportProvider::new(
-                    "http://api.astarte.localhost/pairing".parse().unwrap(),
-                    "secret".to_string(),
-                    None,
-                    true,
-                ),
+                transport_provider,
                 self::connection::Connected,
             ),
             MqttRetention::new(ret_rx),
@@ -986,7 +991,7 @@ pub(crate) mod test {
             .returning(|_, _, _, _| notify_success());
 
         let (mut client, _mqtt_connection) =
-            mock_mqtt_connection(client, eventl, MemoryStore::new());
+            mock_mqtt_connection(client, eventl, MemoryStore::new()).await;
 
         client
             .extend_interfaces(&interfaces, &to_add)
@@ -1040,7 +1045,8 @@ pub(crate) mod test {
             })
             .returning(|_, _, _, _| notify_success());
 
-        let (mut client, _connection) = mock_mqtt_connection(client, eventl, MemoryStore::new());
+        let (mut client, _connection) =
+            mock_mqtt_connection(client, eventl, MemoryStore::new()).await;
 
         client
             .extend_interfaces(&interfaces, &to_add)
@@ -1102,7 +1108,7 @@ pub(crate) mod test {
             });
 
         let (mut mqtt_client, _mqtt_connection) =
-            mock_mqtt_connection(client, eventl, MemoryStore::new());
+            mock_mqtt_connection(client, eventl, MemoryStore::new()).await;
 
         mqtt_client
             .extend_interfaces(&interfaces, &to_add)